rename 'serial' --> 'fserial'
